+++ README.md
@@ -1,64 +1,16 @@
-# MyAngularApp
 
-This project was generated using the [Angular CLI](https://github.com/angular/angular-cli) version 19.1.3.
 
-A foundational **DEX "web-template"** project designed for rapid development and seamless integration of components.
-
-## **Current Development Features**
-
-- **Core Navigation System:**
-  - Fully implemented routing for page transitions and modal windows to ensure smooth and intuitive navigation.
-
-- **Global State Management:**
-  - Utilizing **NgRx** for a scalable and reactive state management system that enhances data flow and consistency throughout the application.
-
-- **Minimalist Design Theme:**
-  - Monochromatic base design with an accent color, supporting dark/light themes for an optimal user experience.
-
-- **User Preferences Module:**
-  - A dedicated **user-preferences component** that centralizes and manages user-specific settings through the global store, ensuring persistent and customized experiences.
-
-## **Development Commands**
-
-### **Serve Application:**
-To run a development server locally, use:
 ```bash
+# clone the repo
+git clone https://github.com/vapaee/telos-underground-defi-website.git telos-underground
+cd telos-underground
+# use the correct npm version to install and run it
+nvm use v22.12.0;
+npm i
 npx ng serve
 ```
-This will launch the project on localhost for demonstration and testing purposes.
 
-### **Project Information:**
-To quickly access key project information and core files, run:
+Fast
 ```bash
-./show_project.sh
+nvm use v22.12.0; rm .angular/ -fr; npx ng serve
 ```
-This script simplifies the process of copying relevant project details during development.
-
----
-
-## **Monochromatic Design with Accent Colors: Why It Matters for DEX**
-
-The **monochromatic design with an accent color**, implemented using **CSS variables**, is central to delivering a professional, clear, and user-friendly interface in this DEX project.
-
-### **Key Benefits:**
-
-1. **Visual Simplicity and Clarity:**
-   - A monochromatic base, consisting of shades of black, white, and gray, minimizes distractions and allows users to focus on essential information.
-
-2. **Effective User Guidance:**
-   - The single accent color serves as a visual cue for key interactions such as buttons, links, and alerts, enhancing user navigation and decision-making.
-
-3. **Dynamic Theme Adaptability:**
-   - Using **CSS variables**, the system efficiently supports both dark and light themes by dynamically adjusting core colors without the need for complex overrides.
-
-4. **Consistency and Scalability:**
-   - This approach ensures that all UI components maintain a unified design language, making the project scalable while reducing maintenance overhead.
-
-5. **Enhanced Visual Communication:**
-   - For a **DEX platform**, where users rely on rapid data interpretation and transactional accuracy, this design maximizes contrast and readability while directing attention to key elements.
-
-By leveraging **CSS variables** and this design paradigm, the project achieves a balance between aesthetic minimalism and functional precision—a critical combination for any decentralized exchange or financial platform.
-
-
-
-
