+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "my-angular-app",
-  "version": "0.0.0",
+  "name": "telos-underground-defi-website",
+  "version": "0.1.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
