+++ src/app/app.component.scss
@@ -1,0 +1,13 @@
+:host {
+    display: block;
+    margin: 0 auto;
+}
+
+.view {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    border: 1px dashed var(--c3);
+}
+
