+++ src/app/app.routes.ts
@@ -1,8 +1,8 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from '@app/core/pages/home.component';
-import { TradeComponent } from '@app/core/pages/trade.component';
-import { ExploreComponent } from '@app/core/pages/explore.component';
-import { PoolComponent } from '@app/core/pages/pool.component';
+import { HomeComponent } from '@app/pages/home/home.component';
+import { TradeComponent } from '@app/pages/trade/trade.component';
+import { ExploreComponent } from '@app/pages/explore/explore.component';
+import { PoolComponent } from '@app/pages/pool/pool.component';
 
 export const routes: Routes =  [
   // Navigate to 'home' component by default (empty path)
