+++ src/app/components/nav-bar/nav-bar.component.html
@@ -1,0 +1,68 @@
+<nav>
+    <!-- Left side: DEX logo + Hamburger Icon (visible on small screens) -->
+    <div class="nav-left">
+        <a 
+            routerLink="/" 
+            routerLinkActive="active-link" 
+            [routerLinkActiveOptions]="{ exact: true }"
+        >
+        <svg width="40" height="40" viewBox="0 0 100 100" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <defs>
+            <linearGradient id="gradientStroke" x1="0%" y1="0%" x2="100%" y2="0%">
+                <stop offset="0%" stop-color="var(--c3)" />
+                <stop offset="100%" stop-color="var(--cc3)" />
+            </linearGradient>
+            </defs>
+            <circle cx="50" cy="50" r="40" stroke="url(#gradientStroke)" stroke-width="12" fill="none" />
+        </svg>
+
+        </a>
+        <a 
+            class="hamburger" 
+            (click)="toggleMenu()" 
+            aria-label="Toggle navigation menu" 
+            [attr.aria-expanded]="isMenuOpen"
+            [attr.aria-controls]="menuId"
+        >
+            &#9776;
+        </a>
+
+        <!-- Center: Nav links (visible on large screens) -->
+        <div class="nav-center">
+            <a routerLink="/trade" routerLinkActive="active-link">Trade</a>
+            <a routerLink="/explore" routerLinkActive="active-link">Explore</a>
+            <a routerLink="/pool" routerLinkActive="active-link">Pool</a>
+        </div>
+    </div>
+
+    <!-- Right side: always visible -->
+    <div class="nav-right">
+        <a class="setting-icon" (click)="toggleSettings()"> &#9881; </a>
+        <a class="search-icon" (click)="toggleSearch()"> &#9906; </a>
+        <a class="btn" (click)="toggleWallet()">Wallet</a>
+    </div>
+</nav>
+
+<!-- Window Container for Navigation Links (visible on small screens when hamburger is clicked) -->
+<app-window-container *ngIf="isMenuOpen" (close)="closeMenu()">
+    <div class="menu-links">
+        <a routerLink="/trade" routerLinkActive="active-link" (click)="closeMenu()">Trade</a>
+        <a routerLink="/explore" routerLinkActive="active-link" (click)="closeMenu()">Explore</a>
+        <a routerLink="/pool" routerLinkActive="active-link" (click)="closeMenu()">Pool</a>
+    </div>
+</app-window-container>
+
+<!-- Window Container for Settings -->
+<app-window-container *ngIf="isSettingsOpen" (close)="closeSettings()">
+    <app-user-preferences></app-user-preferences>
+</app-window-container>
+
+<!-- Window Container for Search -->
+<app-window-container *ngIf="isSearchOpen" (close)="closeSearch()">
+    <p>Search component here</p>
+</app-window-container>
+
+<!-- Window Container for My Wallet -->
+<app-window-container *ngIf="isWalletOpen" (close)="closeWallet()">
+    <p>My Wallet content here</p>
+</app-window-container>
