+++ src/app/components/nav-bar/nav-bar.component.scss
@@ -1,0 +1,69 @@
+nav {
+    padding: 1rem 2rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    background-color: var(--c0);
+    position: relative; /* for absolutely-positioned nav links */
+}
+
+/* Left side: Logo + Hamburger in a row */
+.nav-left {
+    display: flex;
+    align-items: center;
+    gap: 0.3rem;
+}
+
+/* Center: Nav links for large screens */
+.nav-center {
+    display: flex;
+    gap: 1rem;
+}
+
+/* Right side: */
+.nav-right {
+    display: flex;
+    align-items: center;
+    gap: 0.3rem;
+}
+
+/* Hamburger icon for smaller screens */
+.hamburger {
+    display: none;    /* hidden by default, shown in mobile */
+    font-size: 1.3rem;
+    cursor: pointer;
+}
+.setting-icon, .search-icon {
+    font-size: 1.5rem;
+    cursor: pointer;
+}
+
+/* Active link style */
+.active-link {
+    color: var(--c5);
+}
+
+/* Menu Links Inside Window Container */
+.menu-links {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+  /* -------- Responsive Rules -------- */
+@media (max-width: 600px) {
+    /* Hide center nav links on small screens */
+    .nav-center {
+        display: none;
+    }
+
+    /* Show hamburger on small screens */
+    .hamburger {
+        display: block;
+    }
+
+    /* Optionally, adjust nav padding */
+    nav {
+        padding: 0.5rem 1rem;
+    }
+}
