+++ src/app/components/user-preferences/user-preferences.component.html
@@ -1,0 +1,36 @@
+<div class="preferences-container">
+    <h2>User Preferences</h2>
+    
+    <!-- Toggle Theme Button -->
+    <button class="btn" (click)="toggleTheme()">
+        Switch to {{ (isDarkTheme$ | async) ? 'Light' : 'Dark' }} Theme
+    </button>
+    
+    <!-- Hue Slider -->
+    <div class="input-group">
+        <label for="hue">Hue (--h): {{ (hue$ | async) }}</label>
+        <input 
+            type="range" 
+            id="hue" 
+            [value]="(hue$ | async)" 
+            (input)="updateHue($event)" 
+            min="0" 
+            max="360" 
+        />
+        <div class="color-display" [style.backgroundColor]="'hsl(' + (hue$ | async) + ', 50%, 50%)'"></div>
+    </div>
+    
+    <!-- Hue1 Slider -->
+    <div class="input-group">
+        <label for="hue1">Hue1 (--h1): {{ (hue1$ | async) }}</label>
+        <input 
+            type="range" 
+            id="hue1" 
+            [value]="(hue1$ | async)" 
+            (input)="updateHue1($event)" 
+            min="0" 
+            max="360" 
+        />
+        <div class="color-display" [style.backgroundColor]="'hsl(' + (hue1$ | async) + ', 50%, 50%)'"></div>
+    </div>
+</div>
