+++ src/app/components/user-preferences/user-preferences.component.scss
@@ -1,0 +1,28 @@
+.preferences-container {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    padding: 20px;
+    border-radius: 10px;
+}
+
+.input-group {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.input-group label {
+    width: 100px;
+}
+
+.input-group input[type="range"] {
+    flex: 1;
+}
+
+.color-display {
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+    border: 1px solid #ccc;
+}
