--- src/app/core/pages/explore.component.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  standalone: true,
-  selector: 'app-explore',
-  imports: [],
-  template: `
-      <h2>Explore Page</h2>
-      <p>Discover new content here!</p>
-  `,
-  styles: [`
-
-  `]
-})
-export class ExploreComponent {}
