--- src/app/core/pages/pool.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  standalone: true,
-  selector: 'app-trade',
-  template: `
-    <h2>Pool Page</h2>
-  `,
-  styles: [`
-
-  `]
-})
-export class PoolComponent {}
