--- src/app/core/pages/trade.component.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  standalone: true,
-  selector: 'app-trade',
-  imports: [],
-  template: `
-    <h2>Trade Page</h2>
-    <p>Perform your trades here!</p>
-  `,
-  styles: [`
-
-  `]
-})
-export class TradeComponent {}
