--- src/app/core/store/index_effects.ts
@@ -1,14 +1,0 @@
-// src/app/core/store/index_effects.ts
-
-//import { UserPreferencesEffects } from './user_preferences/user-preferences.effects';
-// Import additional effects here as you create them
-// import { WalletInfoEffects } from './wallet_info/wallet-info.effects';
-// import { DataFetchEffects } from './data_fetch/data-fetch.effects';
-// ... other imports
-
-export const AppEffects = [
-  //UserPreferencesEffects,
-  // WalletInfoEffects,
-  // DataFetchEffects,
-  // ... other effects
-];
