+++ src/app/pages/explore/explore.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+
+@Component({
+    standalone: true,
+    selector: 'app-explore',
+    imports: [],
+    template: `
+        <h2>Explore Page</h2>
+        <p>Discover new content here!</p>
+    `,
+    styles: [`
+
+    `]
+})
+export class ExploreComponent {}
