+++ src/app/store/user/index.ts
@@ -1,0 +1,7 @@
+import { actions } from './user.actions';
+import { selectors } from './user.selectors';
+
+export const user = {
+    actions,
+    selectors,
+};
