--- src/styles.css
@@ -1,89 +1,0 @@
-/* styles.css */
-
-/* Default Dark Theme */
-:root {
-  --h: 170;
-  --h1: 215;
-
-  --c0: hsl(var(--h), 0%, 6%); /* dark */
-  --c1: hsl(var(--h), 0%, 8%); /* dark-b */
-  --c2: hsl(var(--h), 30%, 8%); /* accent-b */
-  --c3: hsl(var(--h), 56%, 35%); /* accent */
-  --c4: hsl(var(--h), 0%, 60%); /* light-b */
-  --c5: hsl(var(--h), 10%, 95%); /* light */
-
-  --cc2: hsl(var(--h1), 30%, 8%); /* accent-b */
-  --cc3: hsl(var(--h1), 56%, 35%); /* accent */
-}
-
-/* Dark Theme */
-body.dark-theme {
-  --c0: hsl(var(--h), 0%, 6%); /* dark */
-  --c1: hsl(var(--h), 0%, 9%); /* dark-b */
-  --c2: hsl(var(--h), 30%, 8%); /* accent-b */
-  --c3: hsl(var(--h), 56%, 35%); /* accent */
-  --c4: hsl(var(--h), 0%, 60%); /* light-b */
-  --c5: hsl(var(--h), 0%, 95%); /* light */
-
-  --cc2: hsl(var(--h1), 30%, 8%); /* accent-b */
-  --cc3: hsl(var(--h1), 56%, 35%); /* accent */
-}
-
-/* Light Theme */
-body.light-theme {
-  --c0: hsl(var(--h), 0%, 86%); /* light */ 
-  --c1: hsl(var(--h), 0%, 75%); /* light-b */
-  --c2: hsl(var(--h), 50%, 70%); /* accent-b */
-  --c3: hsl(var(--h), 50%, 50%); /* accent */
-  --c4: hsl(var(--h), 0%, 0%); /* dark */
-  --c5: hsl(var(--h), 0%, 35%); /* dark-b */
-
-  --cc2: hsl(var(--h1), 30%, 8%); /* accent-b */
-  --cc3: hsl(var(--h1), 56%, 35%); /* accent */
-}
-
-/* Global base styles here */
-body {
-  margin: 0;
-  font-family: sans-serif;
-  color: var(--c4);
-  background-color: var(--c0);
-}
-
-.contained{
-border: 1px dashed var(--c4);
-border-radius: 20px;
-padding: 10px 20px;
-}
-.btn{
-font-size: 10pt;
-padding: 10px 20px;
-color: var(--c0);
-border: 1px solid var(--c3);
-background-color: var(--c3);
-border-radius: 20px;
-transition: 0.2s;
-}
-.btn:hover{
-background-color: var(--c0);
-cursor: pointer;
-color: var(--c3);
-}
-a {
-text-decoration: none;
-margin: 10px;
-color: var(--c4);
-}
-a:hover {
-color: var(--c3);
-}
-
-input[type="range"] {
-  -webkit-appearance: none;
-  width: 100%;
-  height: 8px;
-  border-radius: 5px;
-  background: var(--c1);
-  outline: none;
-  margin: 0;
-}
